pyproject.toml
[project]
name = "ai-research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ddgs>=9.11.2",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.135.1",
    "langchain>=1.2.10",
    "langgraph>=1.0.10",
    "openai>=2.26.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]
